requirements.txt
asgiref==3.3.4
Django==3.2.3
django-tinymce4-lite==1.8.0
django-versatileimagefield==2.0
# jsmin==2.2.2
# Pillow==8.2.0
python-decouple==3.4
python-magic==0.4.22
pytz==2021.1
sqlparse==0.4.1
